Fix multiple code quality and correctness issues
- NWC: persist wallet connection (remember_wallet=true), add create_dir_all for secure storage, fix cfg guards with compile_error!, add preserve_storage parameter to disconnect_nwc
- V4V: clear error before new send attempt in CustomBoostInput
- Voice: add disabled state for native platforms in message_card
- Wiki: hoist ToastOptions/Duration imports to module level
- Android: trim/filter signer packages, validate signed event payload
- Git: use url crate for proper URL redaction
- GitHub: validate NIP numbers (hex format, 2-3 chars) before fetching
- Bookmarks: use AtomicU32 instead of AtomicU64 for generation counter
- Relay: gate storage import with #[cfg(feature = "web")]
- Pinned: fix cfg mismatch, use runtime feature guard
- ICS: preserve TZID in LocalDateTime export
- Tests: fix unused import warning in timed_serializer

+++ src/stores/bookmarks.rs
@@ -55,7 +55,7 @@ pub static BOOKMARK_SYNC_STATUS: GlobalSignal<BookmarkSyncStatus> = Signal::glob
 pub static BOOKMARK_ROLLBACK_STATE: GlobalSignal<Store<BookmarkRollbackStore>> = Signal::global(||
 Store::new(BookmarkRollbackStore::default()));
 /// Generation counter to track bookmark changes and prevent stale publishes
-static BOOKMARK_GENERATION: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
+static BOOKMARK_GENERATION: std::sync::atomic::AtomicU32 = std::sync::atomic::AtomicU32::new(0);
 #[cfg(feature = "web")]
 thread_local! {
     /// Pending bookmark publish timeout (for debouncing)
@@ -212,7 +212,7 @@ pub async fn unbookmark_event(event_id: String) -> Result<(), String> {
 fn publish_with_retry(
     bookmarks: Vec<String>,
     retry_count: u32,
-    generation: u64,
+    generation: u32,
 ) -> std::pin::Pin<Box<dyn std::future::Future<Output = ()> + Send + 'static>> {
     Box::pin(async move {
         // Check if this publish is still current (not stale)
@@ -283,7 +283,7 @@ fn publish_with_retry(
 fn publish_with_retry(
     bookmarks: Vec<String>,
     retry_count: u32,
-    generation: u64,
+    generation: u32,
 ) -> std::pin::Pin<Box<dyn std::future::Future<Output = ()> + 'static>> {
     Box::pin(async move {
         // Check if this publish is still current (not stale)
